+++ frontend/src/app/components/home/coupons/coupon/coupon-details/coupon-details.component.ts
@@ -6,7 +6,7 @@ import {
 } from '@angular/common';
 import { Component, inject, Input, OnInit, Signal } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
-import { CouponDto } from '../../../../../../dtos/coupon.dto';
+import { CouponDto, UpdateCouponDto } from '../../../../../../dtos/coupon.dto';
 import { MatButtonModule } from '@angular/material/button';
 import { MatMenuModule } from '@angular/material/menu';
 import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
@@ -17,6 +17,10 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { CustomDatePipe } from '../../../../../pipe/date.pipe';
 import { DeleteDialogComponent } from '../../../common/delete-dialog/delete-dialog.component';
 import { CouponStore, OnCouponSuccess } from '../store/coupon.store';
+import { UserAbility, UserAbilityTuple } from '../../../../../permissions/ability';
+import { PureAbility } from '@casl/ability';
+import { AbilityServiceSignal } from '@casl/angular';
+import { NotAllowedDialogBoxComponent } from '../../../../common/not-allowed-dialog-box/not-allowed-dialog-box.component';
 
 @Component({
   selector: 'app-coupon-details',
@@ -43,6 +47,10 @@ export class CouponDetailsComponent implements OnInit {
 
   organization = this.organizationStore.organizaiton;
 
+  private readonly abilityService = inject<AbilityServiceSignal<UserAbility>>(AbilityServiceSignal);
+	protected readonly can = this.abilityService.can;
+	private readonly ability = inject<PureAbility<UserAbilityTuple>>(PureAbility);
+
   constructor(private router: Router, private route: ActivatedRoute) {}
 
   ngOnInit(): void {
@@ -59,32 +67,50 @@ export class CouponDetailsComponent implements OnInit {
   }
 
   openDialog(coupon: CouponDto) {
-    this.dialog.open(DeleteDialogComponent, {
-      data: {
-        title: `Delete ‘${coupon.name}’ coupon?`,
-        description: `Are you sure you want to delete ‘${coupon.name}’? All campaigns and coupon codes associated with this coupon will be deleted!`,
-        onDelete: () =>
-          this.couponStore.deleteCoupon({
-            organizationId: this.organization()?.organizationId!,
-            couponId: coupon.couponId!,
-          }),
-      },
-      autoFocus: false,
-    });
+    if(this.can('delete', CouponDto)) {
+      this.dialog.open(DeleteDialogComponent, {
+        data: {
+          title: `Delete ‘${coupon.name}’ coupon?`,
+          description: `Are you sure you want to delete ‘${coupon.name}’? All campaigns and coupon codes associated with this coupon will be deleted!`,
+          onDelete: () =>
+            this.couponStore.deleteCoupon({
+              organizationId: this.organization()?.organizationId!,
+              couponId: coupon.couponId!,
+            }),
+        },
+        autoFocus: false,
+      });
+    } else {
+      const rule = this.ability.relevantRuleFor('delete', CouponDto);
+      this.openNotAllowedDialogBox(rule?.reason!);
+    }
   }
 
   onEdit() {
-    this.router.navigate(
-      [
-        `/${this.organization()?.organizationId}/coupons/${
-          this.coupon()?.couponId
-        }/edit`,
-      ],
-      {
-        queryParams: {
-          redirect: btoa(this.router.url),
-        },
-      }
-    );
+    if(this.can('update', UpdateCouponDto)) {
+      this.router.navigate(
+        [
+          `/${this.organization()?.organizationId}/coupons/${
+            this.coupon()?.couponId
+          }/edit`,
+        ],
+        {
+          queryParams: {
+            redirect: btoa(this.router.url),
+          },
+        }
+      );
+    } else {
+      const rule = this.ability.relevantRuleFor('update', UpdateCouponDto);
+      this.openNotAllowedDialogBox(rule?.reason!);
+    }
   }
+
+  openNotAllowedDialogBox(restrictionReason: string) {
+		this.dialog.open(NotAllowedDialogBoxComponent, {
+			data: {
+				description: restrictionReason,
+			}
+		});
+	}
 }
